requirements.txt
# Añadir estas dependencias
google-cloud-storage==2.10.0
django-storages==1.14.2 